vstart: add public_network option for cephadm usage
This commit introduces the ability to specify the public_network option
when using cephadm. Previously, the public_network was automatically
selected based on the output of the ip route list command.
However, this approach posed a problem when the command failed to make
a selection (either due to the absence of routes or the presence of
multiple routes), leading to a failure in the ceph config set operation.

With this update, users now have the flexibility to explicitly define
the public_network option during cephadm deployment, enhancing
reliability and control in network configurations.

Signed-off-by: daegon.yang <daegon.yang@samsung.com>